Arm refactoring.
I refactored the arms to be a map instead of a vector. This allows for more
intuitive (and probably faster) arm update and removal from a bank of arms.

One side effect is that some operations now return a tuple. The most
significant of these are the algorithm specific select-arm's. Two functions
have been added to hide this detail. See tuple-arm and tuple-idx.