fixed ucb1 bug; simplified simulation-seq->table
There was an order of operations bug in the UCB1 implementation. I was stumped
as to why the graphs looked slightly off from what I expected and tracked it
down to this bug.

I also simplified the simulation-seq->table fn by moving the simulation number
into simulation-seq. I think it makes the most sense there because it is a
"label" on each row of a given simulation seq. It also makes
simulation->seq->table much simpler.