Using score instead of value in ucb1.
I was updating the value attribute of the arm with the curiosity bonus. I
wasn't comfortable with this since the returned arm might later be manipulated
with a value that didn't make any sense. To address this, I update the "score"
which won't follow the arm (and doesn't need to) in the map of arms.