Kitchen sink commit.
This commit has been lingering for days. It includes several minor tweaks to
bring mab up-to-date some of the different assumptions we have in the casino.

One fix is to randomly pick an arm in epsilon greedy when all of the values are
0. This allows the algorithm to get around always picking one or other of the
ends (depending on the length of the arms) when we know nothing about them.

The other addition is to add a step parameter to update-arm. The assumption
that was main when building update arm is that the counts would always
increment by one. That's not always the case.